Fix performance issue with item quantity
When loading inventory, there was a large wait time for the page to
load due to the app making many times more requests to the database
than it needed to. This was because quantity was calculated with a
count for every item. I have changed it so that there is a single group
by count done in inventory.py that is then passed to templates cutting
the number of requests hugely