Inventory performance (#18)
* Fix performance issue with item quantity

When loading inventory, there was a large wait time for the page to
load due to the app making many times more requests to the database
than it needed to. This was because quantity was calculated with a
count for every item. I have changed it so that there is a single group
by count done in inventory.py that is then passed to templates cutting
the number of requests hugely

* Fix crash on non-admin

* Add test for item quantities on inventory page

This is to check the correctness of ff47e6

* Fix item quantities on inventory page

Before it was returning how many items hadn't been released in a
fulfilled request, but the behaviour before was that it would tell how
many hadn't been approved. The new behaviour is the latter